+++ src/enums/documentEncoding.ts
@@ -1,3 +1,0 @@
-export enum DocumentEncoding {
-	utf8 = 'UTF-8'
-}