+++ src/index.ts
@@ -1,13 +1,13 @@
-import './extensions/elements/txt';
 import './extensions/elements/baseElements';
 import './extensions/elements/svgElements';
 import './extensions/attributes/baseAttributes';
 import './extensions/attributes/ariaAttributes';
 import './extensions/attributes/svgAttributes';
 import './extensions/attributes/eventAttributes';
-import './extensions/overlap';
 import './extensions/components';
-import './extensions/foreach'
+import './extensions/overlap';
+import './extensions/logic';
+import './extensions/txt';
 
 import { Tag } from "./tag";